refactor(src): breaking circular dependencies
breaking circular dependencies between GraphData, and HasData

BREAKING CHANGE: hasDataMethod to hasData - "method" naming is implicit  and unnecessary
