+++ src/components/bcg-matrix-chart/bcg-matrix-chart.tsx
@@ -111,7 +111,7 @@ export class BGCMatrixChart implements Graph<BcgMatrix[]> {
   }
 
   hasData(): Error | boolean {
-    return this.graphDataMerged.hasDataMethod(this.graphDataMerged);
+    return this.graphDataMerged.hasData(this.graphDataMerged);
   }
 
   reSetRoot() {
