+++ src/components/horizontal-bar-chart/horizontal-bar-chart.tsx
@@ -94,7 +94,7 @@ export class HorizontalBarChart implements Graph<number[]> {
   }
 
   hasData(): Error | boolean {
-    return this.graphDataMerged.hasDataMethod(this.graphDataMerged);
+    return this.graphDataMerged.hasData(this.graphDataMerged);
   }
 
   reSetRoot(): void {
