+++ src/components/line-annotations-chart/line-annotations-chart.tsx
@@ -98,7 +98,7 @@ export class LineAnnotationsChart implements Graph {
   }
 
   hasData(): Error | boolean {
-    return this.graphDataMerged.hasDataMethod(this.graphDataMerged);
+    return this.graphDataMerged.hasData(this.graphDataMerged);
   }
 
   reSetRoot(): void {
