+++ src/interfaces/graph-meta.ts
@@ -1,4 +1,4 @@
-import { GraphData } from './index';
+import { GraphData } from './graph-data';
 
 export type GraphMeta<T = number[][]> = {
   width: number;
