--- src/interfaces/has-data-strategy.ts
@@ -1,6 +1,0 @@
-import { GraphData } from './../interfaces';
-
-export type HasDataStrategy<T = number[][]> = (
-  graphData: GraphData<T>,
-  data?: T,
-) => Error | boolean;
