+++ src/interfaces/has-data.ts
@@ -1,0 +1,1 @@
+export type HasData<T = unknown> = (data: T) => Error | boolean;
