+++ src/shared/index.ts
@@ -1,5 +1,3 @@
-import { Formats } from './formats';
-
 import {
   DEFAULT_GRAPH_DATA_PIE,
   DEFAULT_GRAPH_DATA_BAR,
@@ -10,7 +8,6 @@ import {
 } from './default-graph-data';
 
 export {
-  Formats,
   DEFAULT_GRAPH_DATA_PIE,
   DEFAULT_GRAPH_DATA_BAR,
   DEFAULT_GRAPH_DATA_LINE,
